mr

See merge request ai_track/class_08/ai-project/team02/team2!16

+++ back/src/services/commentService.ts
@@ -14,7 +14,7 @@ const prisma = new PrismaClient();
 export async function createdComment(
   inputData: {
     content: string;
-    nestedComment: string;
+    nestedComment: string; // 댓글(null)인지 대댓글(원댓글의 id)인지 확인
   },
   authorId: string,
   diary_id: string,
@@ -69,14 +69,13 @@ export async function getCommentByDiaryId(
   limit: number,
 ) {
   try {
-    // 댓글 작성자의 id, username, porfileImage를 함께 응답
-    // 대댓글은 reComment에 배열로 포함하여 응답
     const comment = await prisma.comment.findMany({
       skip: (page - 1) * limit,
       take: limit,
       where: { diaryId: diary_id, nestedComment: null },
       select: {
         id: true,
+        // 댓글 작성자의 id, username, porfileImage를 함께 응답
         author: {
           select: {
             id: true,
@@ -88,6 +87,7 @@ export async function getCommentByDiaryId(
         content: true,
         createdAt: true,
         updatedAt: true,
+        // 대댓글은 reComment에 배열로 포함하여 응답
         reComment: {
           select: {
             id: true,
@@ -108,6 +108,7 @@ export async function getCommentByDiaryId(
       orderBy: { createdAt: 'asc' },
     });
 
+    // 댓글이 없을 경우 응답
     if (comment.length == 0) {
       const response = emptyApiResponseDTO();
       return response;
@@ -170,21 +171,29 @@ export async function updatedComment(
       emojis[Math.floor(Math.random() * emojis.length)];
     const emoji = randomEmoji.emotion;
 
-    const comment = await prisma.comment.update({
-      where: { id: comment_id, authorId },
-      data: {
-        content: inputData.content,
-        emoji: emoji,
-      },
+    // 댓글 작성자 본인인지 확인을 위한 조회
+    const userCheck = await prisma.comment.findUnique({
+      where: { id: comment_id },
     });
 
-    const commentResponseData = plainToClass(commentResponseDTO, comment, {
-      excludeExtraneousValues: true,
-    });
+    // 댓글 작성자가 맞다면 수정 진행
+    if (userCheck.authorId == authorId) {
+      const comment = await prisma.comment.update({
+        where: { id: comment_id },
+        data: {
+          content: inputData.content,
+          emoji: emoji,
+        },
+      });
 
-    const response = successApiResponseDTO(commentResponseData);
+      const commentResponseData = plainToClass(commentResponseDTO, comment, {
+        excludeExtraneousValues: true,
+      });
 
-    return response;
+      const response = successApiResponseDTO(commentResponseData);
+
+      return response;
+    }
   } catch (error) {
     throw error;
   }
@@ -193,17 +202,25 @@ export async function updatedComment(
 // 댓글 삭제
 export async function deletedComment(comment_id: string, authorId: string) {
   try {
-    const comment = await prisma.comment.delete({
-      where: { id: comment_id, authorId },
+    // 댓글 작성자 본인인지 확인을 위한 조회
+    const userCheck = await prisma.comment.findUnique({
+      where: { id: comment_id },
     });
 
-    const commentResponseData = plainToClass(commentResponseDTO, comment, {
-      excludeExtraneousValues: true,
-    });
+    // 댓글 작성자가 맞다면 삭제 진행
+    if (userCheck.authorId == authorId) {
+      const comment = await prisma.comment.delete({
+        where: { id: comment_id },
+      });
 
-    const response = successApiResponseDTO(commentResponseData);
+      const commentResponseData = plainToClass(commentResponseDTO, comment, {
+        excludeExtraneousValues: true,
+      });
 
-    return response;
+      const response = successApiResponseDTO(commentResponseData);
+
+      return response;
+    }
   } catch (error) {
     throw error;
   }
